Added the License to the Project
In this internal release, we introduce significant changes to the project by adding a license for the first time. This addition clearly defines the terms and conditions for the use of the software, providing a solid legal foundation. The chosen license, Apache License 2.0, reflects our commitment to transparency and open collaboration. This step is crucial to establish clear guidelines on how others can contribute, use, and distribute the software in compliance. We appreciate all contributors for their dedication, and we hope that this inclusion fosters a more structured and accessible development environment for everyone.